requirements.txt
google-cloud-pubsub>=2.13.0
colorama>=0.4.6
python-dotenv>=1.0.0
fastapi>=0.115.0
starlette>=0.41.3,<0.47.0
uvicorn>=0.23.0
jinja2>=3.1.2
websockets>=11.0.3 